Implement Dynamic Checkout Button Remove native overlay on WebView Add loading indicator with cancel button
This implementation will download the .svg file from the remote file server and render it
as a PNG. CheckoutButtonManager will handle this downloading and configuring of the
button and its image.

Removing the native overlay and introducing a loading indicator. This will provide the user
the option to cancel the transaction in the event the WebView does not load. The
cancel button provided by the page will suffice once it's loaded.